Began shape integration
Line complete, as yet untested.
Then do other shapes.
Then texturing
Then text w/ add in bulk!

+++ pmma/core/cpp_src/Animation/LinearAnimation.cpp
@@ -2,7 +2,7 @@
 
 #include "PMMA_Core.hpp"
 
-CPP_LinearAnimation::CPP_LinearAnimation(CPP_DisplayCoordinate* NewTargetCoordinatePtr) {
+CPP_LinearAnimation::CPP_LinearAnimation(CPP_DisplayCoordinate *NewTargetCoordinatePtr) {
     TargetCoordinatePtr = NewTargetCoordinatePtr;
 
     StartCoordinatePtr = new CPP_DisplayCoordinate();
@@ -46,7 +46,7 @@ void CPP_LinearAnimation::Start() {
     StartTime = std::chrono::high_resolution_clock::now();
     RunTime = std::chrono::seconds(0);
 
-    float start_position[2];
+    uint16_t start_position[2];
     StartCoordinatePtr->Get(start_position);
     TargetCoordinatePtr->Set(start_position);
 }
